requirements.txt
bcrypt==4.3.0
boto3==1.37.4
botocore==1.37.4
bounded-pool-executor==0.0.3
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
cryptography==44.0.2
dotenv==0.9.9
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-core==2.4.2
google-cloud-storage==3.1.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.68.0
idna==3.10
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jmespath==1.0.1
keyring==25.6.0
more-itertools==10.6.0
mysql-connector-python==9.2.0
paramiko==3.5.1
pqdm==0.2.0
prettytable==3.15.1
proto-plus==1.26.0
protobuf==5.29.3
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyNaCl==1.5.0
PyQt6==6.8.1
PyQt6-Qt6==6.8.2
PyQt6_sip==13.10.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pywin32-ctypes==0.2.3
requests==2.32.3
rsa==4.9
s3transfer==0.11.3
schedule==1.2.2
setuptools==75.8.2
six==1.17.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
wcwidth==0.2.13
